Create Comparison of Quick Sort, Merge Sort and Radix Sort
Perform the same implementation, with additional functions for quick and
radix sort. Compare the performance of these 3 sorting algorithms (merge, quick, radix) on an
array of 10000 random integers. Perform the simulation 100 times, each time initialising new
values in the array. Then compare the results in terms of best, worst, and average times for
each algorithm.