Create rearranging a number of large crates in order of the time they are to be shipped out.cpp
A clerk at a shipping company is charged with the task of rearranging a number of large
crates in order of the time they are to be shipped out. Thus, the cost of compares is very low
relative to the cost of exchanges (move the crates). The warehouse is nearly full: there is
extra space sufficient to hold any one of the crates, but not two. Which sorting method should
the clerk use? Implement this question via a user generated array?